Replaced int with uint8 where appropriate (#5)

+++ src/automation-hat/adc.go
@@ -2,7 +2,7 @@ package automation_hat
 
 // 3.3V ADC, with no LED
 type adc struct {
-	channel    int
+	channel    uint8
 	maxVoltage float64
 }
 
@@ -12,10 +12,10 @@ type adcLed struct {
 	ledContainer
 }
 
-func GetADC33(channel int, maxVoltage float64) ADC33 {
+func GetADC33(channel uint8, maxVoltage float64) ADC33 {
 	return adc{channel: channel, maxVoltage: maxVoltage}
 }
 
-func GetADC24(channel int, maxVoltage float64, ledPin int) ADC24 {
+func GetADC24(channel uint8, maxVoltage float64, ledPin uint8) ADC24 {
 	return adcLed{adc: adc{channel: channel, maxVoltage: maxVoltage}, ledContainer: ledContainer{GetLED(ledPin)}}
 }
